move cli skeleton under src/
