+++ src/cmd/checkaccess.go
