+++ src/cmd/validate.go
