+++ src/main.go
@@ -3,7 +3,7 @@ Copyright © 2024 NAME HERE <EMAIL ADDRESS>
 */
 package main
 
-import "github.com/osusec/beavercds-ng/cmd"
+import "github.com/osusec/beavercds-ng/src/cmd"
 
 func main() {
 	cmd.Execute()
